discord-bot/requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
pyyaml>=6.0
playwright>=1.40.0
Pillow>=10.0.0

